+++ IBAnimatable/PresentationAnimationType.swift
@@ -53,3 +53,13 @@ public enum PresentationAnimationType: IBEnum {
   }
 
 }
+
+extension PresentationAnimationType: Hashable {
+  public var hashValue: Int {
+    return stringValue.hashValue
+  }
+  
+  public static func == (lhs: PresentationAnimationType, rhs: PresentationAnimationType) -> Bool {
+    return lhs.stringValue == rhs.stringValue
+  }
+}