+++ IBAnimatable/TransitionAnimationType.swift
@@ -241,3 +241,13 @@ extension TransitionAnimationType {
     case rotate
   }
 }
+
+extension TransitionAnimationType: Hashable {
+  public var hashValue: Int {
+    return stringValue.hashValue
+  }
+  
+  public static func == (lhs: TransitionAnimationType, rhs: TransitionAnimationType) -> Bool {
+    return lhs.stringValue == rhs.stringValue
+  }
+}