+++ CHANGELOG.md
@@ -11,6 +11,8 @@ None
 #### Enhancements
 - Add support for corner and shadow on UICollectionViewCell.
 [#368]https://github.com/IBAnimatable/IBAnimatable/pull/368
+- Use Swift Enum as key in our internal presentation and transition cache.
+[#360](https://github.com/IBAnimatable/IBAnimatable/issues/360)
 
 #### Bugfixes
 - Fix initial `cornerSide` rendering issue.